test: cover portal nav item builder for help tab
