+++ apps/web/src/components/public/portal-header.tsx
@@ -11,6 +11,7 @@ import {
   HelpCenterCategoryNav,
   type HelpCenterCategory,
 } from '@/components/help-center/help-center-category-nav'
+import { buildNavItems } from './portal-header-nav'
 import { useIntl, FormattedMessage } from 'react-intl'
 import { cn } from '@/lib/shared/utils'
 import { isTeamMember } from '@/lib/shared/roles'
@@ -52,42 +53,6 @@ interface PortalHeaderProps {
   }
   /** Whether to show the theme toggle (hidden when admin forces a specific theme) */
   showThemeToggle?: boolean
-}
-
-const NAV_ITEMS_BASE = [
-  { to: '/', messageId: 'portal.header.nav.feedback', defaultMessage: 'Feedback' },
-  { to: '/roadmap', messageId: 'portal.header.nav.roadmap', defaultMessage: 'Roadmap' },
-  { to: '/changelog', messageId: 'portal.header.nav.changelog', defaultMessage: 'Changelog' },
-] as const
-
-const NAV_ITEM_HELP = {
-  to: '/hc',
-  messageId: 'portal.header.nav.help',
-  defaultMessage: 'Help',
-} as const
-
-export type PortalNavItem = (typeof NAV_ITEMS_BASE)[number] | typeof NAV_ITEM_HELP
-
-/**
- * Returns the nav items shown in the portal header.
- * On the help center subdomain we show only the Help tab so the standalone
- * experience stays focused. Elsewhere we show feedback/roadmap/changelog and
- * append a Help tab when the help center is enabled.
- */
-export function buildNavItems({
-  helpCenterEnabled,
-  helpCenterHost,
-}: {
-  helpCenterEnabled: boolean
-  helpCenterHost: boolean
-}): readonly PortalNavItem[] {
-  if (helpCenterHost) {
-    return helpCenterEnabled ? [NAV_ITEM_HELP] : []
-  }
-  if (helpCenterEnabled) {
-    return [...NAV_ITEMS_BASE, NAV_ITEM_HELP]
-  }
-  return NAV_ITEMS_BASE
 }
 
 export function PortalHeader({
